requirements.txt
absl-py==0.6.1
astor==0.7.1
backports.weakref==1.0.post1
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.1.post0
grpcio==1.17.1
h5py==2.9.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
mock==2.0.0
numpy==1.15.4
pbr==5.1.1
protobuf==3.6.1
scipy==1.2.0
six==1.12.0
tensorboard==1.11.0
tensorflow-gpu==1.11.0
termcolor==1.1.0
tialpath==0.1.2
Werkzeug==0.14.1
